Fix GMT offset conversion for fractional timezones
Previously, using the  function or casting to  to convert the GMT offset to a fractional decimal number could cause regression problems for fractional timezones. Now, we use the  function to correctly convert the GMT offset to a decimal number. This resolves the issue of incorrect conversion of fractional timezones.
fixes: https://core.trac.wordpress.org/ticket/58986
props marcio-zebedeu