WIP: Rely on Unicode processing within a RegExp pattern for title sanitization.
From time to time new issues arise with the sanitization of a post title for slug creation.
The existing algorithm builds takes a hit-and-miss approach of handling specific cases of
known string elements that cause problems and replacing them with normalized characters.

In this patch a given title is first converted into a normalized form and then processed
with a Unicode-aware PCRE pattern which formalizes the _kinds_ of replacements which are
supposed to occur. For example, instead of removing "%c2%ab" the code now removes
"invisible characters" as defined by Unicode itself.

This update, if it works without breaking existing dependencies, poses a more comprehensive
solution to the problem of slug generation, one that updates with advancements to the
Unicode specification provided by system libraries and PHP itself instead of through
custom WordPress code.